Moving network attr init feature to the agent
